refactor: move templates to a better package
- Also moving output package for public so addons can also print debug
  messages.
