+++ internal/template/api/common.tmpl
