+++ internal/template/api/domain.tmpl
