+++ internal/template/api/embed.go
