+++ internal/template/api/enum.tmpl
