+++ internal/template/api/outbound.tmpl
