+++ internal/template/api/routes.tmpl
