+++ internal/template/api/validation.tmpl
