+++ internal/template/api/wire.tmpl
