+++ internal/template/api/wire_input.tmpl
