+++ internal/template/testing/embed.go
